replace `Logger` with `BaseLogger`, and make new `Logger` class user-aware
`base_logger` files are just renamed from `logger` files
