+++ libs/logging/src/export.ts
@@ -10,7 +10,7 @@ import { JsonStreamInput, jsonStream } from '@votingworks/utils';
 import { z } from 'zod';
 import { LogEventId } from './log_event_ids';
 import { CLIENT_SIDE_LOG_SOURCES } from './base_types/log_source';
-import { type Logger } from './logger';
+import { type BaseLogger } from './base_logger';
 import { DEVICE_TYPES_FOR_APP, LogLineSchema, LoggingUserRole } from './types';
 
 function extractAdditionalKeysFromObj(
@@ -27,7 +27,7 @@ function extractAdditionalKeysFromObj(
 }
 
 async function* generateCdfEventsForExport(
-  logger: Logger,
+  logger: BaseLogger,
   currentUser: LoggingUserRole,
   logFileReader: AsyncIterable<string>
 ): AsyncGenerator<EventLogging.Event> {
@@ -81,7 +81,7 @@ async function* generateCdfEventsForExport(
 }
 
 export async function* buildCdfLog(
-  logger: Logger,
+  logger: BaseLogger,
   electionDefinition: ElectionDefinition,
   logFileReader: AsyncIterable<string>,
   machineId: string,
