+++ libs/logging/src/index.ts
@@ -3,6 +3,7 @@ export * from './base_types';
 export * from './export';
 export * from './helpers';
 export * from './log_event_ids';
-export * from './logger';
 export * from './test_utils';
 export * from './types';
+export * from './base_logger';
+export * from './logger';
